+++ vtpfrontend/src/views/TradePage/StockList.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Box from '@mui/material/Box';
 import { DataGrid } from '@mui/x-data-grid';
-import TopNavBar from '../../components/TopNavBar';
+import TopNavBar from '../../components/TopNavbar/TopNavBar';
 import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
